Initial commit, with openclipart.py
openclipart.py contains a utility to quickly download and filter free
clipart from openclipart.org. It uses openclipart.org's API to find the
most-downloaded images, then downloads a lower-res version of each of
them and shows them to the user. If the user selects "y" to use the
image, it downloads the vector version of the image and places it in the
graphics/ directory.